backend/requirements.txt
Django==4.1.1
django-crontab==0.7.1
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mysqlclient==2.1.1
packaging==21.3
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.2.1
redis==4.3.4
sqlparse==0.4.2
wrapt==1.14.1